Move `PathShapes` to partial class
